:construction: WIP: refacto implm
